Refactored folder structure, removed /getlyrics command, now you can search directly for song name without any specific command
